fix: use custom meta `max-width` to override `prettierOptions.printWidth`
- Extracted `parseMetaString` function to `src/utils/parseMetaString.ts`
